Create Stopwatch web application
I have created codes to build a stopwatch web application, using HTML, CSS, and JavaScript. HTML is used to structure the elements of the application. By Implementing functions for starting, pausing, and resetting the stopwatch, as well as tracking and displaying lap times, so that users can accurately measure and record  time intervals.